+++ eos_invoices/__init__.py
@@ -1,4 +1,4 @@
 """Outstanding payments of a Corporation across Alliance Auth apps."""
 
-__version__ = "0.0.1"
+__version__ = "0.0.2"
 VERSION = __version__